El emulador habla SDL3 directo, y la ventana de referencia siempre fue un FBO
Portado de SDL 1.2 (via sdl12-compat -> sdl2-compat -> SDL3, tres DLL) a la API
de SDL3 3.4.16: main.c (ventana, eventos, la redireccion de stdout/stderr junto
al exe que hacia SDLmain), graficos.c (SDL_CreateWindow + SDL_GL_CreateContext),
audio.c (SDL_OpenAudioDeviceStream), hilo.c (SDL_Mutex/SDL_Condition), BFont.c y
debug.c (superficies). main.h ya no incluye SDL: tests/ compila sin cabeceras de
SDL. CMake descarga el paquete oficial con SHA256 fijado y copia una sola DLL.

Dos reglas del port: el SDL_opengl.h de SDL3 se saltea entero, glext incluido,
si <GL/gl.h> vino antes (graficos.c incluye SDL_opengl_glext.h a mano); y SDL3
repite teclas, que SDL 1.2 no hacia, asi que se descarta event.key.repeat.

Lo que el port destapo: la primera compuerta dio Crazy Taxi con 3714 pixeles a
+-1 LSB y la ejecucion identica al digito. sdl12-compat dibujaba en un FBO
propio de 800x600 y lo escalaba a la ventana, asi que la referencia de siempre
era un FBO y SDL3 directo fue el primero en rasterizar de verdad en la ventana,
donde el driver de AMD redondea distinto. --render=ventana rasteriza ahora en un
FBO del tamano de la ventana copiado 1:1 (destino_ventana) y los tres guests
vuelven byte a byte a build-ref; DCEMU_VENTANA_DIRECTA=1 es el A/B.

Medido antes de adoptarse: compuerta de juegos byte a byte contra build-ref en
SR2, DCDoom y CT (captura, wav, puntos, entregas; con y sin --hilos), parque KOS
151/151 sin una diferencia, y la tanda de dos binarios (herramientas/
binarios-ab.ps1, nueva) contra el arbol anterior reentrenado en un worktree:
DCDoom -0,9 %, SR2 -1,9 %, CT -7,8 %, los tres con rangos disjuntos y 4/4,
totales al digito. La ganancia sobrevive a --sin-audio (CT -9,2 %, disjuntos):
es el camino del cuadro y los eventos, no el del audio.

+++ traza.c
@@ -10,6 +10,9 @@
 
 /* Para desensamblar el bucle donde se trabo y volcar los registros. traza.c
    solo se enlaza en el emulador, nunca en las pruebas. */
+/* SDL3 ya no llega por main.h: aqui es por SDL_GetTicks(). */
+#include <SDL3/SDL.h>
+
 #include "main.h"
 #include "debug.h"
 #include "sh4emu.h"
@@ -705,7 +708,7 @@ void traza_resumen(void)
 	   que compararlo. Ver docs/clock-plan.md, fase 4. */
 	{
 		unsigned long long emulado = reloj_ms();
-		unsigned long real = SDL_GetTicks();
+		unsigned long real = (unsigned long) SDL_GetTicks();
 
 		if (emulado > 0 && real > 0)
 			fprintf(stderr, "traza: %llu ms emulados en %lu ms reales (%.2fx)\n",
